[error] add receipt error
