+++ cita-chain/types/src/errors/mod.rs
@@ -1,5 +1,7 @@
 mod error;
 mod execution;
+mod receipt;
 
 pub use error::Error;
 pub use execution::ExecutionError;
+pub use receipt::ReceiptError;
